requirements.txt
google-genai>=0.8.0
streamlit>=1.24
langchain>=0.0.200
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
PyMuPDF>=1.22.0         # fitz
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=9.5.0
python-dotenv>=1.0.0
openai>=0.27.0
tqdm>=4.65.0
pandas>=2.0.0
typing_extensions

# LangChain + Google GenAI
langchain
langchain-core
langchain-community
google-genai
google-generativeai
langchain-google-genai

# Testing & PDF generation
pytest>=7.0
reportlab>=3.6.12

